Switch REST Client implementation to RESTEasy
- Updates to RESTEasy 4.1.1.Final
- Removes SmallRye REST Client extension and implementation
- Adds REST Client extension using RESTEasy
- Fixes #3055
