+++ bom/runtime/pom.xml
@@ -16,7 +16,7 @@
     <properties>
         <aesh.version>1.11</aesh.version>
         <jandex.version>2.1.1.Final</jandex.version>
-        <resteasy.version>4.1.0.Final</resteasy.version>
+        <resteasy.version>4.1.1.Final</resteasy.version>
         <opentracing.version>0.31.0</opentracing.version>
         <opentracing-jaxrs.version>0.4.1</opentracing-jaxrs.version>
         <opentracing-web-servlet-filter.version>0.2.3</opentracing-web-servlet-filter.version>
@@ -29,7 +29,7 @@
         <microprofile-config-api.version>1.3</microprofile-config-api.version>
         <microprofile-opentracing-api.version>1.3.1</microprofile-opentracing-api.version>
         <microprofile-reactive-streams-operators.version>1.0</microprofile-reactive-streams-operators.version>
-        <microprofile-rest-client.version>1.2.1</microprofile-rest-client.version>
+        <microprofile-rest-client.version>1.3.3</microprofile-rest-client.version>
         <smallrye-config.version>1.3.5</smallrye-config.version>
         <smallrye-health.version>1.0.3</smallrye-health.version>
         <smallrye-metrics.version>1.1.3</smallrye-metrics.version>
@@ -41,7 +41,6 @@
         <smallrye-reactive-streams-operators.version>1.0.6</smallrye-reactive-streams-operators.version>
         <smallrye-converter-api.version>1.0.4</smallrye-converter-api.version>
         <smallrye-reactive-messaging.version>0.0.10</smallrye-reactive-messaging.version>
-        <smallrye-rest-client.version>1.2.2</smallrye-rest-client.version>
         <swagger-ui.version>3.20.9</swagger-ui.version>
         <javax.activation.version>1.1.1</javax.activation.version>
         <javax.inject.version>1</javax.inject.version>
@@ -444,7 +443,7 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
-                <artifactId>quarkus-smallrye-rest-client</artifactId>
+                <artifactId>quarkus-rest-client</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -1093,29 +1092,6 @@
                 </exclusions>
             </dependency>
             <dependency>
-                <groupId>io.smallrye</groupId>
-                <artifactId>smallrye-rest-client</artifactId>
-                <version>${smallrye-rest-client.version}</version>
-                <exclusions>
-                    <exclusion>
-                        <groupId>org.eclipse.microprofile.config</groupId>
-                        <artifactId>microprofile-config-api</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>io.smallrye</groupId>
-                        <artifactId>smallrye-config</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>org.jboss.resteasy</groupId>
-                        <artifactId>resteasy-client</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>org.jboss.resteasy</groupId>
-                        <artifactId>resteasy-json-p-provider</artifactId>
-                    </exclusion>
-                </exclusions>
-            </dependency>
-            <dependency>
                 <groupId>javax.activation</groupId>
                 <artifactId>activation</artifactId>
                 <version>${javax.activation.version}</version>
@@ -1496,6 +1472,17 @@
                     <exclusion>
                         <groupId>javax.activation</groupId>
                         <artifactId>activation</artifactId>
+                    </exclusion>
+                </exclusions>
+            </dependency>
+            <dependency>
+                <groupId>org.jboss.resteasy</groupId>
+                <artifactId>resteasy-client-microprofile</artifactId>
+                <version>${resteasy.version}</version>
+                <exclusions>
+                    <exclusion>
+                        <groupId>org.eclipse.microprofile.config</groupId>
+                        <artifactId>microprofile-config-api</artifactId>
                     </exclusion>
                 </exclusions>
             </dependency>
