+++ devtools/common/src/main/filtered/extensions.json
@@ -532,15 +532,14 @@
     "guide": "https://quarkus.io/guides/amqp-guide"
   },
   {
-    "name": "SmallRye REST Client",
+    "name": "REST Client",
     "labels": [
-      "smallrye-rest-client",
       "rest-client",
       "web-client",
       "microprofile-rest-client"
     ],
     "groupId": "io.quarkus",
-    "artifactId": "quarkus-smallrye-rest-client",
+    "artifactId": "quarkus-rest-client",
     "guide": "https://quarkus.io/guides/rest-client-guide"
   },
   {
