+++ docs/src/main/asciidoc/rest-client-guide.adoc
@@ -42,7 +42,7 @@ mvn io.quarkus:quarkus-maven-plugin:{quarkus-version}:create \
     -Dextensions="rest-client, resteasy-jsonb"
 ----
 
-This command generates the Maven project with a REST endpoint and imports the `smallrye-rest-client` and `resteasy-jsonb` extensions.
+This command generates the Maven project with a REST endpoint and imports the `rest-client` and `resteasy-jsonb` extensions.
 
 
 == Setting up the model
