+++ extensions/rest-client/runtime/src/main/java/io/quarkus/restclient/runtime/BuilderResolver.java
@@ -1,0 +1,15 @@
+package io.quarkus.restclient.runtime;
+
+import org.eclipse.microprofile.rest.client.RestClientBuilder;
+import org.eclipse.microprofile.rest.client.spi.RestClientBuilderResolver;
+import org.jboss.resteasy.microprofile.client.RestClientBuilderImpl;
+
+/**
+ * Created by hbraun on 22.01.18.
+ */
+public class BuilderResolver extends RestClientBuilderResolver {
+    @Override
+    public RestClientBuilder newBuilder() {
+        return new RestClientBuilderImpl();
+    }
+}
