+++ extensions/rest-client/runtime/src/main/java/io/quarkus/restclient/runtime/graal/ClientBuilderReplacement.java
@@ -1,0 +1,17 @@
+package io.quarkus.restclient.runtime.graal;
+
+import javax.ws.rs.client.ClientBuilder;
+
+import org.jboss.resteasy.client.jaxrs.internal.ResteasyClientBuilderImpl;
+
+import com.oracle.svm.core.annotate.Substitute;
+import com.oracle.svm.core.annotate.TargetClass;
+
+@TargetClass(ClientBuilder.class)
+final class ClientBuilderReplacement {
+
+    @Substitute
+    public static ClientBuilder newBuilder() {
+        return new ResteasyClientBuilderImpl();
+    }
+}
