+++ extensions/rest-client/runtime/src/main/resources/META-INF/services/org.jboss.resteasy.microprofile.client.header.IncomingHeadersProvider
@@ -1,0 +1,1 @@
+io.quarkus.restclient.runtime.IncomingHeadersProvider
