--- extensions/smallrye-rest-client/runtime/src/main/resources/META-INF/services/io.smallrye.restclient.header.IncomingHeadersProvider
@@ -1,1 +1,0 @@
-io.quarkus.smallrye.restclient.runtime.IncomingHeadersProvider
