+++ integration-tests/main/pom.xml
@@ -103,7 +103,7 @@
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-smallrye-rest-client</artifactId>
+            <artifactId>quarkus-rest-client</artifactId>
         </dependency>
 
         <!-- JPA -->
