+++ tcks/microprofile-rest-client/tck-suite.xml
@@ -102,11 +102,6 @@
                     <exclude name=".*" />
                 </methods>
             </class>
-            <class name="org.eclipse.microprofile.rest.client.tck.ClientHeaderParamTest">
-                <methods>
-                    <exclude name=".*" />
-                </methods>
-            </class>
             <class name="org.eclipse.microprofile.rest.client.tck.ClientReuseTest">
                 <methods>
                     <exclude name=".*" />
